+++ src/main/java/taketengaming/tencore/TenCore.java
@@ -2,7 +2,6 @@ package taketengaming.tencore;
 
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.Mod.EventHandler;
-import net.minecraftforge.fml.common.event.FMLInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLServerStartingEvent;
 import taketengaming.tencore.register.Commands;
 
@@ -12,7 +11,7 @@ public class TenCore
 	public static final String MODID = "tencore";
 	public static final String NAME = "TenCore";
 	public static final String UPDATEURL = "https://raw.githubusercontent.com/TakeTenGaming/TenCore/master/versions.json";
-	public static final String VERSION = "1.1.0";
+	public static final String VERSION = "1.2.0";
 
 	@Mod.Instance
 	public static TenCore instance;
@@ -20,15 +19,8 @@ public class TenCore
 	public static Logger logger = new Logger ( "TenCore" );
 
 	@EventHandler
-	public void init ( FMLInitializationEvent event )
-	{
-		logger.debug ( "Heeeelllooo Minecraft!" );
-	}
-
-	@EventHandler
 	public void serverStarting ( FMLServerStartingEvent event )
 	{
-		logger.debug ( "Server starting.." );
 		Commands.serverStarting ( event );
 	}
 }