+++ src/main/java/taketengaming/tencore/item/ItemArmor.java
@@ -15,11 +15,12 @@ public class ItemArmor extends net.minecraft.item.ItemArmor
 	{
 		super ( material, 0, equipmentSlot );
 
-		name = NameUtil.getNameLegacy ( name );
+		name = NameUtil.getName ( name );
 		this.setRegistryName ( name );
 		this.setUnlocalizedName ( name );
 	}
 
+	@SuppressWarnings( "ConstantConditions" )
 	public ModelResourceLocation getModelResourceLocation ()
 	{
 		return new ModelResourceLocation ( this.getRegistryName (), "inventory" );