+++ src/main/java/taketengaming/tencore/item/ItemAxe.java
@@ -12,11 +12,12 @@ public class ItemAxe extends net.minecraft.item.ItemAxe
 	{
 		super ( material, material.getDamageVsEntity (), material.getEfficiencyOnProperMaterial () );
 
-		name = NameUtil.getNameLegacy ( name );
+		name = NameUtil.getName ( name );
 		this.setRegistryName ( name );
 		this.setUnlocalizedName ( name );
 	}
 
+	@SuppressWarnings( "ConstantConditions" )
 	public ModelResourceLocation getModelResourceLocation ()
 	{
 		return new ModelResourceLocation ( this.getRegistryName (), "inventory" );