+++ src/main/java/taketengaming/tencore/item/ItemPickaxe.java
@@ -12,11 +12,12 @@ public class ItemPickaxe extends net.minecraft.item.ItemPickaxe
 	{
 		super ( material );
 
-		name = NameUtil.getNameLegacy ( name );
+		name = NameUtil.getName ( name );
 		this.setRegistryName ( name );
 		this.setUnlocalizedName ( name );
 	}
 
+	@SuppressWarnings( "ConstantConditions" )
 	public ModelResourceLocation getModelResourceLocation ()
 	{
 		return new ModelResourceLocation ( this.getRegistryName (), "inventory" );