+++ src/main/java/taketengaming/tencore/item/ItemShovel.java
@@ -14,11 +14,12 @@ public class ItemShovel extends ItemSpade
 	{
 		super ( material );
 
-		name = NameUtil.getNameLegacy ( name );
+		name = NameUtil.getName ( name );
 		this.setRegistryName ( name );
 		this.setUnlocalizedName ( name );
 	}
 
+	@SuppressWarnings( "ConstantConditions" )
 	public ModelResourceLocation getModelResourceLocation ()
 	{
 		return new ModelResourceLocation ( this.getRegistryName (), "inventory" );