+++ src/main/java/taketengaming/tencore/recipes/BaseRecipeHandler.java
@@ -1,84 +1,87 @@
 package taketengaming.tencore.recipes;
 
+import com.google.common.collect.Maps;
 import net.minecraft.block.Block;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 
-import java.util.ArrayList;
-import java.util.Collection;
+import java.util.Map;
 
 /**
  * Created by Acid on 2/1/2017.
  */
 public class BaseRecipeHandler
 {
-	private ArrayList< Recipe > recipes = new ArrayList<> ();
+	public static final BaseRecipeHandler instance = new BaseRecipeHandler ();
+	public Map< ItemStack, ItemStack > recipes = Maps.newHashMap ();
 
 	public void addRecipe ( ItemStack input, ItemStack output )
 	{
-		this.recipes.add ( new Recipe ( input, output ) );
+		this.recipes.put ( input, output );
 	}
 
 	public void addRecipe ( Block input, Block output )
 	{
-		this.recipes.add ( new Recipe ( input, output ) );
+		this.recipes.put ( new ItemStack ( input ), new ItemStack ( output ) );
 	}
 
 	public void addRecipe ( Block input, Item output )
 	{
-		this.recipes.add ( new Recipe ( input, output ) );
+		this.recipes.put ( new ItemStack ( input ), new ItemStack ( output ) );
 	}
 
 	public void addRecipe ( Block input, ItemStack output )
 	{
-		this.recipes.add ( new Recipe ( input, output ) );
+		this.recipes.put ( new ItemStack ( input ), output );
 	}
 
 	public void addRecipe ( Item input, Item output )
 	{
-		this.recipes.add ( new Recipe ( input, output ) );
+		this.recipes.put ( new ItemStack ( input ), new ItemStack ( output ) );
 	}
 
-	public void addRecipes ( Collection recipes )
+	public void addRecipes ( Map< ItemStack, ItemStack > recipes )
 	{
-		this.recipes.addAll ( recipes );
+		this.recipes.putAll ( recipes );
 	}
 
-	protected boolean compareItemStacks ( ItemStack stack1, ItemStack stack2 )
+	public static BaseRecipeHandler getInstance ()
 	{
-		return stack2.getItem () == stack1.getItem () && ( stack2.getMetadata () == 32767 || stack2.getMetadata () == stack1.getMetadata () );
+		return instance;
 	}
 
-	public boolean containsRecipe ( Recipe recipe )
+	public ItemStack getRecipe ( ItemStack itemStack )
 	{
-		return this.recipes.contains ( recipe );
-	}
+		ItemStack testStack = itemStack.copy ();
 
-	public Recipe getRecipe ( ItemStack itemStack )
-	{
-		for ( Recipe recipe : this.getRecipes () )
+		for ( Map.Entry< ItemStack, ItemStack > entry : this.recipes.entrySet () )
 		{
-			if ( this.compareItemStacks ( itemStack, recipe.getInput () ) )
+			if ( itemStack.getCount () > entry.getKey ().getCount () )
 			{
-				return recipe;
+				testStack.setCount ( entry.getKey ().getCount () );
+			}
+
+			if ( ItemStack.areItemsEqual ( testStack, entry.getKey () ) )
+			{
+				return entry.getValue ();
 			}
 		}
 
-		return null;
+		return ItemStack.EMPTY;
 	}
 
-	public ArrayList< Recipe > getRecipes ()
+	public Map< ItemStack, ItemStack > getRecipes ()
 	{
 		return this.recipes;
 	}
 
-	public void removeRecipe ( Recipe recipe )
+	public boolean hasRecipe ( ItemStack itemStack )
 	{
-		if ( !this.containsRecipe ( recipe ) )
-		{
-			return;
-		}
+		return this.recipes.containsKey ( itemStack );
+	}
 
-		this.recipes.remove ( this.recipes.indexOf ( recipe ) );
+	public void removeRecipe ( ItemStack itemStack )
+	{
+		this.recipes.remove ( itemStack );
 	}
 }